Define firmware, persist and cache mount points for sdxpoorwills
This commit declares /firmware, /cache and /persist to be treated
as mount point for mounting modem partition, cache and persist
volume respectively. It also decouples selinux from mounting
logic. During build it is detected that selinux is enabled or
not. If it is not enabled then selinux rule is removed from
mounting partition logic.

Change-Id: I0a74ac85884a572b185f1918ea7cd01e5598ae17