Increase rootfs/system volume's size
We need to account for on-the-fly compression in ubifs
which is not as good as the compression achieved at time
of ubifs image creation; so post-OTA-upgrade, while files
are created on-the-fly, there will be less free-space
available. Allocate more volume size to system to accomodate
this.

Also reduce cache volume size since we use /data for storing
OTA update package.

Change-Id: I05a107c8056ff917f0db555dbe783c864c2b5b31