conf: set LAYERSERIES_COMPAT to thud
The LAYERSERIES_COMPAT variable is meant to indicate the list of
compatible version(s) of OE-core. This branch of meta-qti-bsp is
compatible with thud only.

OE-core will check the value of LAYERSERIES_COMPATH with
LAYERSERIES_COMPAT_core, and warn in case an incompatbility is
detected.

As such using LAYERSERIES_COMPAT_core is a hack. While it looks
convenient , it will be misguiding for any user of meta-qti-bsp since
the layer version check will *never* fail.

Change-Id: Ic18b7e7b579e306c2b950124ac8c4481d86ce687
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>