linux-msm: copy image file in do_deploy
we currently copy the zImage in DEPLOY_DIR_IMAGE inside
do_shared_workdir() and this file is used by do_make_bootimg in the
qimage class, and we needed the following dependency:

do_make_bootimg[depends]  += "virtual/kernel:do_shared_workdir"

Because of that whenever we build an image, we need to run
do_shared_workdir first, which means that building from a fresh TMPDIR
folder doesn't work fine and rebuilds many things.

We fix this issue by copying the files in DEPLOYDIR from do_deploy()
which has a valid sstate object, and remove the extra dependency to
help fix this inefficiency. Note that we use DEPLOYDIR and not
DEPLOY_DIR_IMAGE, as per Yocto Project reference manual:

==
When inheriting the deploy class, the DEPLOYDIR points to a temporary
work area for deployed files that is set in the deploy class as
follows:

     DEPLOYDIR = "${WORKDIR}/deploy-${PN}"

Recipes inheriting the deploy class should copy files to be deployed
into DEPLOYDIR, and the class will take care of copying them into
DEPLOY_DIR_IMAGE afterwards.
==

We need to adjust tasks dependencies accordingly, since
do_make_bootimg() now depends on virtual/kernel:do_deploy().

Change-Id: I029789e531484f8d66b60aed199165db0998758d
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>