qdl: add recipe
This is an import for the following 3 commits from:
http://git.yoctoproject.org/cgit/cgit.cgi/meta-qcom/

14164d9 qdl: remove depedency on udev-native for native build
3364427 qdl: Fix QA Issue: No GNU_HASH in the ELF binary
7ce246b qdl: needs pkgconfig and udev
b6b8e0f qdl: add recipe

QDL is a USB flashing tool that implements QTI Sahara and
Firehose protocol, it can be used with a large number of Snapdragon
chipsets.

Note that udev-native does not exist, qdl depends on udev, so to build
qdl-native libudev-dev package needs to be installed on the host
system.

Change-Id: Ie1e23dceecfbd6fbcc3b2551ec89cabee862e59b
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>