qimage: dm-verity: check returned code for subprocess.call
We need to check the return code for subprocess.call(). If the command
fails we need to fail the task and report that to bitbake.

With the current implementation if mkbootimg fails, it would be silent
and the build process would continue.

With this patch, we can detect errors like that:

ERROR: machine-image-1.0-r0 do_make_bootimg: <CMD> failed.

Change-Id: Ia9f5f812d0ab096272d7c13c950ff287de2efd4f
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>