systemd-machine-units: change firmware.mount to wants for local-fs.target
the firmware.mount when installed as wants for local-fs.target does not block
the boot-up when the modem partition is not flashed with a filesystem.

mounting mmcblk0p1 is removed from firmware-links.sh as this is supposed
to be taken care by firmware.mount. also it is not safe to assume that
mmcblk0p1 has modem firmware filesystem

Change-Id: I0ccc0c33c45470d31d3205cb26b532dcc26bf25b