meta-qti-bsp: Make ramdisk creation sstate compatible
Ramdisk creation needs binaries from STAGING_KERNEL_BUILDDIR.
This directory is created as part of do_shared_workdir task.
But this task is not sstate compatible.

So, copy the binaries that are needed to create ramdisk to
deploy directory which is sstate directory. Use the binaries
from this directory to generate ramdisk.

Change-Id: I40de93c1e3cb6169b7c1352f95d64b872d8715ac