gobject-introspectation cleanup
These files seem to be artifacts from the past. Most likely from
Jethro times. In Rocko gobject-introspection was largely fixed in
OE-Core, and none of what we have in meta-qti-bsp is needed now.

The upstream commits for gobject-introspection fix are the following:

7bd32b9ce643 (recipes-gnome: fix introspection support)
efd37c5d3982 (python-pygobject: update to 3.18.2)
03cd714efc8a (libsoup-2.4: enable gobject introspection)

In rocko and beyond it is enough to inherit gobject-introspection, and
the OE core class does the right thing. It supports -native , -target
and sdk.

Change-Id: I8b6728bc778f1f39d13bbcb92e308155f82320e5
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>