qimage: deploy gen_partition artifacts in image specific folder
The gen_partition artifact are specific to each image build, since
they include the image file name and partition size. So far we are
deploying the artifact in DEPLOY_DIR_IMAGE, so if we build several
images, they will be overriden.

This patch makes the following changes:
* deploy these files in DEPLOY_DIR_IMAGE/<image name>
* rely on do_image_complete sstate, by using IMGDEPLOYDIR, it's a good
practice to use sstate as much as possible. That means the task now
depends on do_rootfs which creates IMGDEPLOYDIR.
* fix the permission for partition.xml (which had +x)

Change-Id: Icd8627911bedcf970b59e351e07a236e4ea8ef6b
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>