qimage: do_makesystem: use sstate, and create per image verity folder
Instead of deploying artifact in DEPLOY_DIR_IMAGE, we can use
IMGDEPLOYDIR which is the 'sstate' deploy folder for
do_image_complete.

Using sstate will ensure that files are removed/replaced during the
build, and if files are overriden sstate will detect and fail.

When dm-verity is enabled it creates files in
DEPLOY_DIR_IMAGE/verity. However these files are different for each
image, and so they would be overriden when building multiple
images. In this patch we move the verity artifact in
DEPLOY_DIR_IMAGE/<image name>/verity.

Change-Id: I5d0ac5bccf96113702286916880c1f6aae2a5cbc
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>