Fix Error: Task do_packagedata does not exist for target on eSDK
Recipes that are not named as -native are not considered as native
recipes during build sanity checks. Hence native recipes that are
not referred to as -native during "devtool build" fail on eSDK with
Error: Task do_packagedata does not exist for target. Renaming such
recipes with <recipe>-native.bb fixes the issue.

Change-Id: If5823b7c01ea86117fba9f0a2224adb6f6863434