meta-qti-bsp: Add ttyGS0 node to console list in cinder
Add ttyGS0 to the the list of output consoles in cmdline. TtyGS node is
added before ttyMSM, so that ttyMSM will be chosen as the default system
console (/dev/console). Else ttyGS will chosen as the default console,
hence we wont be able to log into UART shell which will make the UART
shell inaccessible. Also added ttyGS0 to list of SERIAL_CONSOLES, this
will be used by systemd-serialgetty.bb to create agetty services for
each tty node.

Change-Id: Idb8d0b9981a333e680c1210d5f00a0009c86e9e8