Mount sdcard in recovery mode
Earlier automountsdcard.sh script was part of busybox,
which was installing this script under /etc/udev/scripts
directory. Later, this script was moved out of busybox as
few customers were using Toybox.

It looks like the recovery recipe is assuming that the script
will be available by default (via busybox) and hence, the
recovery recipe is not taking care to install the script.
Due to the absence of this script in recovery mode,
possibly the SD Card is not getting mounted.

Change-Id: Ia3854e3f7fc8766ccd6718bdfb4f815ac49e1748
CRs-Fixed: 3250790