meta-qti-bsp: Add MACHINE_FEATURES for ssg bring up of sa415m
1. Added norpmb_listener and nossd_listener because there is
no support for rpmb and ssd. Because sa415m is a NAND target.
2. The sa415m target uses km4 which is compatible with
legcay km_lite. Adding legacy_kmlite as machine feature, so
that module compilation for km_lite can be controlled.
3. The "le_qseecom" machine feature is added so that
gpfspath_le_oem_config.xml is pushed to device. This
xml file is required to add whitelist paths, which is
needed by qseecomd.

Change-Id: I85fce1658c0d825427328a55f665fe93904fa0ab